adding documentation, doc string
